[ENDPOINT] NEW: Now can use SSL encryption on HTTP endpoint
This commit introduces full SSL support for the pistache endpoint.

First of all, in order to enable SSL support in the library, one has to
compile with PISTACHE_USE_SSL option ON. One can then configure the
endpoint to accept / use SSL connections only:

    Http::Endpoint server(addr);
    server.useSSL("./server.crt", "./cert/server.key");

With that done, all the connections to the server shall now be in HTTPS.

One can also enable certificate authentication against the server:

    server.useSSLAuth("./rootCA.crt");

The server will now only accept client connection with a certificate
signed with the Certificate Authority passed in the example above.

Signed-off-by: Ne02ptzero <louis@ne02ptzero.me>

 ________________________________________
/ Q: What do monsters eat? A: Things. Q: \
| What do monsters drink? A: Coke.       |
\ (Because Things go better with Coke.)  /
 ----------------------------------------
        \   ^__^
         \  (oo)\_______
            (__)\       )\/\
                ||----w |
                ||     ||