[LISTENER] FIX: Move useSSL_ out of the PISTACHE_USE_SSL scope
This is required in order not to crash the inline Class initialization
on gcc (Didn't test on clang, behavior should be the same). Long story
short, when you compile the library with PISTACHE_USE_SSL the internal
boolean useSSL_ is used in the Listener class constructor. From the
client, when you include the the library (without #define PISTACHE_USE_SSL),
the allocator will fail horribly on class instancation (Size mismatch
between the compiled library and the client inclusion).

Signed-off-by: Ne02ptzero <louis@ne02ptzero.me>

 _______________________________________
/ Tempt not a desperate man. -- William \
\ Shakespeare, "Romeo and Juliet"       /
 ---------------------------------------
        \   ^__^
         \  (oo)\_______
            (__)\       )\/\
                ||----w |
                ||     ||